Handle result of previous calc as first input.
After the first calculation if the user selects an operator then the
result of the previous calculation should be used as the first number
for the next calculation. This would set num1 to the result and
operator to the entered operator.

If the user types a number after a calculation then the previous result
is discarded and the entry is considered the start of a new calculation.

Previously typing a digit after a caluclation would append the digit to
the previous result.